pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.20.0.post0"
schedule = "^1.2.2"
python-dotenv = "^1.1.0"
loguru = "^0.7.3"
httpx = "^0.28.1"
python-box = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
